# api/swagger.yaml
basePath: /api
definitions:
  internal_controllers.CreateTaskRequest:
    properties:
      performed_at:
        example: "2024-03-20T14:30:00Z"
        type: string
      summary:
        example: Replaced filters and recharged coolant
        type: string
      title:
        example: Fix air conditioning
        type: string
    required:
    - performed_at
    - summary
    - title
    type: object
  internal_controllers.UpdateTaskRequest:
    properties:
      performed_at:
        example: "2024-03-20T14:30:00Z"
        type: string
      summary:
        example: Replaced filters and recharged coolant
        type: string
      title:
        example: Fix air conditioning
        type: string
    required:
    - performed_at
    - summary
    - title
    type: object
  sword-challenge_internal_models.Notification:
    description: Notification information
    properties:
      created_at:
        description: '@Description When the notification was created'
        example: "2024-03-20T14:30:00Z"
        type: string
      id:
        description: '@Description The unique identifier of the notification'
        example: 1
        type: integer
      is_read:
        description: '@Description Whether the notification has been read'
        example: false
        type: boolean
      message:
        description: '@Description The notification message'
        example: The tech John Doe performed the task on 2024-03-20 14:30:00
        type: string
      task_id:
        description: '@Description The ID of the task this notification is about'
        example: 1
        type: integer
    type: object
  sword-challenge_internal_models.Task:
    description: Task information
    properties:
      created_at:
        description: '@Description When the task was created'
        example: "2024-03-20T14:30:00Z"
        type: string
      id:
        description: '@Description The unique identifier of the task'
        example: 1
        type: integer
      performed_at:
        description: '@Description When the task was performed'
        example: "2024-03-20T14:30:00Z"
        type: string
      summary:
        description: '@Description The detailed summary of the task'
        example: Replaced filters and recharged coolant
        type: string
      technician_id:
        description: '@Description The ID of the technician who performed the task'
        example: 1
        type: integer
      title:
        description: '@Description The title of the task'
        example: Fix air conditioning
        type: string
      updated_at:
        description: '@Description When the task was last updated'
        example: "2024-03-20T14:30:00Z"
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A task management API for technicians and managers.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sword Challenge API
  version: "1.0"
paths:
  /api/notifications:
    get:
      consumes:
      - application/json
      description: Get all unread notifications for the authenticated manager
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/sword-challenge_internal_models.Notification'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get unread notifications
      tags:
      - notifications
  /api/notifications/{id}/read:
    put:
      consumes:
      - application/json
      description: Mark a specific notification as read
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark notification as read
      tags:
      - notifications
  /api/tasks:
    get:
      consumes:
      - application/json
      description: Get all tasks for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/sword-challenge_internal_models.Task'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Create a new task for the authenticated technician
      parameters:
      - description: Task Information
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.CreateTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/sword-challenge_internal_models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new task
      tags:
      - tasks
  /api/tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a task
      tags:
      - tasks
    get:
      consumes:
      - application/json
      description: Get a task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sword-challenge_internal_models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a specific task
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Update an existing task
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Task Information
        in: body
        name: task
        required: true
        schema:
          $ref: '#/definitions/internal_controllers.UpdateTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sword-challenge_internal_models.Task'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a task
      tags:
      - tasks
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"